+++ pygce/models/garmin.py
@@ -17,6 +17,7 @@
 
 
 import json
+from datetime import datetime
 
 from bs4 import BeautifulSoup
 
@@ -31,6 +32,17 @@ def parse_num(n):
 
     m = str(n).strip().replace(".", "").replace(",", ".")
     return float(m)
+
+
+def parse_hours_minutes(h):
+    """
+    :param h: str
+        Hours and minutes in the form hh:mm to parse
+    :return: datetime.time
+        Time parsed
+    """
+
+    return datetime.strptime(str(h).strip(), "%H:%M").time()
 
 
 class GCDaySection(object):
@@ -69,7 +81,11 @@ class GCDaySection(object):
             A json representation of this object
         """
 
-        return json.dumps(self.to_dict())
+        d = self.to_dict()
+        for k in d.keys():
+            d[k] = str(d[k])  # convert to string to be json serializable
+
+        return json.dumps(d)
 
 
 class GCDaySummary(GCDaySection):
@@ -209,8 +225,78 @@ class GCDaySleep(GCDaySection):
 
         GCDaySection.__init__(self, raw_html)
 
+        self.night_sleep_time = None
+        self.nap_time = None
+        self.total_sleep_time = None  # typically night_sleep + nap
+        self.bed_time = None
+        self.wake_time = None
+        self.deep_sleep_time = None
+        self.light_sleep_time = None
+        self.awake_sleep_time = None  # time during night you were awake
 
-class GCDayActivity(GCDaySection):
+    def parse(self):
+        self.parse_sleep_totals()
+        self.parse_bed_time()
+        self.parse_sleep_times()
+
+    def parse_sleep_totals(self):
+        """
+        :return: void
+            Finds value of night/nap/total sleep times
+        """
+
+        container = self.soup.find_all("div", {"class": "equation centered"})[0]
+        times = container.find_all("div", {"class": "data-bit"})
+        times = [str(t.text).strip() for t in times]  # strip texts
+
+        self.night_sleep_time = parse_hours_minutes(times[0])
+        self.nap_time = parse_hours_minutes(times[1])
+        self.total_sleep_time = parse_hours_minutes(times[2].split(" ")[0])
+
+    def parse_bed_time(self):
+        """
+        :return: void
+            Finds hour start/end sleep
+        """
+
+        times = self.soup.find_all("div", {"class": "time-inline-edit-placeholder"})
+        times = [str(t.text).strip() for t in times]  # strip texts
+
+        self.bed_time = datetime.strptime(times[0], "%I:%M %p").time()  # account for AM/PM
+        self.wake_time = datetime.strptime(times[1], "%I:%M %p").time()  # account for AM/PM
+
+    def parse_sleep_times(self):
+        """
+        :return: void
+            Finds deep/light/awake sleep times
+        """
+
+        container = self.soup.find_all("div", {
+            "class": "span4 text-center sleep-chart-secondary deep-sleep-circle-chart-placeholder"})[0]
+        self.deep_sleep_time = parse_hours_minutes(container.find_all("span")[0].text.split("hrs")[0])
+
+        container = self.soup.find_all("div", {
+            "class": "span4 text-center sleep-chart-secondary light-sleep-circle-chart-placeholder"})[0]
+        self.light_sleep_time = parse_hours_minutes(container.find_all("span")[0].text.split("hrs")[0])
+
+        container = self.soup.find_all("div", {
+            "class": "span4 text-center sleep-chart-secondary awake-circle-chart-placeholder"})[0]
+        self.awake_sleep_time = parse_hours_minutes(container.find_all("span")[0].text.split("hrs")[0])
+
+    def to_dict(self):
+        return {
+            "night_sleep_time": self.night_sleep_time,
+            "nap_time": self.nap_time,
+            "total_sleep_time": self.total_sleep_time,
+            "bed_time": self.bed_time,
+            "wake_time": self.wake_time,
+            "deep_sleep_time": self.deep_sleep_time,
+            "light_sleep_time": self.light_sleep_time,
+            "awake_sleep_time": self.awake_sleep_time
+        }
+
+
+class GCDayActivities(GCDaySection):
     """
     Standard activity in the Garmin Connect timeline of day.
     Common features are kcal, time, distance, type, name, link
@@ -275,7 +361,7 @@ class GCDayTimeline(object):
             "summary": GCDaySummary(summary_html),
             "steps": GCDaySteps(steps_section_html),
             "sleep": GCDaySleep(sleep_section_html),
-            "activities": GCDayActivity(activities_section_html),
+            "activities": GCDayActivities(activities_section_html),
             "breakdown": GCDayBreakdown(breakdown_section_html)
         }  # list of sections in day